+++ pure_extensions/lib/src/dart/collections/map.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 
-extension MapExtensions<K, V> on Map<K, V> {
-  void removeNullValues() => removeWhere((key, value) => value == null);
+import 'package:pure_extensions/src/dart/collections/iterable.dart';
 
+extension MapExtensions<K, V> on Map<K, V> {
   String encodeToString() => jsonEncode(this);
 
   Iterable<dynamic> serialize() {
@@ -28,7 +28,6 @@ extension MapExtensions<K, V> on Map<K, V> {
   }
 
   /// Returns the first entry that satisfies the given predicate [test].
-  ///
   /// [Iterable.firstWhere]
   MapEntry<K, V> firstWhere(
     bool Function(K key, V value) test, {
@@ -38,7 +37,6 @@ extension MapExtensions<K, V> on Map<K, V> {
   }
 
   /// Returns the last entry that satisfies the given predicate [test].
-  ///
   /// [Iterable.lastWhere]
   MapEntry<K, V> lastWhere(
     bool Function(K key, V value) test, {
@@ -47,27 +45,42 @@ extension MapExtensions<K, V> on Map<K, V> {
     return entries.lastWhere((entry) => test(entry.key, entry.value), orElse: orElse);
   }
 
+  /// Returns the single element that satisfies [test].
+  /// [Iterable.lastWhere]
+  MapEntry<K, V> singleWhere(
+    bool Function(K key, V value) test, {
+    MapEntry<K, V> Function()? orElse,
+  }) {
+    return entries.singleWhere((entry) => test(entry.key, entry.value), orElse: orElse);
+  }
+
+  /// Returns the first entry that satisfies [test] otherwise null.
+  /// [Iterable.firstWhere]
+  MapEntry<K, V>? tryFirstWhere(bool Function(K key, V value) test) {
+    return entries.tryFirstWhere((entry) => test(entry.key, entry.value));
+  }
+
+  /// Returns the last entry that satisfies [test] otherwise null.
+  /// [Iterable.lastWhere]
+  MapEntry<K, V>? tryLastWhere(bool Function(K key, V value) test) {
+    return entries.tryLastWhere((entry) => test(entry.key, entry.value));
+  }
+
+  /// Returns the single element that satisfies [test].
+  /// [Iterable.singleWhere]
+  MapEntry<K, V>? trySingleWhere(bool Function(K key, V value) test) {
+    return entries.trySingleWhere((entry) => test(entry.key, entry.value));
+  }
+
   /// Returns the first entry if it exists otherwise null.
   ///
   /// [Iterable.first]
-  MapEntry<K, V>? get tryFirst {
-    Iterator<MapEntry<K, V>> it = entries.iterator;
-    if (!it.moveNext()) {
-      return null;
-    }
-    return it.current;
-  }
+  MapEntry<K, V>? get firstOrNull => entries.firstOrNull;
 
   /// Returns the last entry if it exists otherwise null.
   ///
   /// [Iterable.last]
-  MapEntry<K, V>? get tryLast {
-    MapEntry<K, V>? result;
-    for (final entry in entries) {
-      result = entry;
-    }
-    return result;
-  }
+  MapEntry<K, V>? get lastOrNull => entries.lastOrNull;
 
   /// Reduces a map to a single value by iteratively combining entries
   /// of the collection using the provided function.
@@ -97,5 +110,18 @@ extension MapExtensions<K, V> on Map<K, V> {
   /// [Iterable.fold]
   T fold<T>(T initialValue, T combine(T previousValue, MapEntry<K, V> entry)) {
     return entries.fold(initialValue, combine);
+  }
+}
+
+extension MapNullExtensions<K, V> on Map<K?, V?> {
+  /// Returns [Map] without null keys and values.
+  Map<K, V> whereNotNull() {
+    final map = <K, V>{};
+    forEach((key, value) {
+      if (key != null && value != null) {
+        map[key] = value;
+      }
+    });
+    return map;
   }
 }